+++ crudcrate-derive/tests/expand/join_filter_sort.expanded.rs
@@ -610,7 +610,7 @@ mod customer__Model {
                                     __jf.operator,
                                     &__jf.value,
                                     __backend,
-                                )
+                                )?
                             }
                             "year" => {
                                 crudcrate::build_filter_expr::<
@@ -622,7 +622,7 @@ mod customer__Model {
                                     __jf.operator,
                                     &__jf.value,
                                     __backend,
-                                )
+                                )?
                             }
                             _ => None,
                         };